sort frontend changes to support array inserts
when array insertions and array element overwrites happen in the same change,
the frontend does an optimization where the original element value doesn't get
set immediately after the insertion, leading to problems --
when we process an ins, we currently assume the ins + set pair was already
processed for the previous element the ins references.

we considered fancy approaches to this but landed on this hack:
just sort the change from the frontend so an ins is always followed by the
corresponding set. seems to work for now.

if we end up wanting to not do this sorting, what we'll need to do is
change our index<->elemId conversion functions to, instead of using
the _elemIds skiplist data structure inside automerge as they do now,
use other structures like _insertion that track pending insertions.